Adjustment to tests execution
