+++ test/functional-api/integration/search/search.it-spec.ts
@@ -1,4 +1,4 @@
-import { getUser } from '@test/functional/user-management/user.request.params';
+import { getUser } from '@test/functional-api/user-management/user.request.params';
 import '@test/utils/array.matcher';
 import {
   createTestEcoverse,
@@ -12,7 +12,7 @@ import {
   hostNameId,
   organisationName,
 } from '../organisation/organisation.request.params';
-import { searchMutation } from '../search/search.request.params';
+import { searchMutation } from './search.request.params';
 
 const userEmail = 'qa.user@alkem.io';
 const userName = 'QA User';
