+++ test/functional-e2e/registration/registration-page-object.ts
